Add More Retries
Sometimes, the AWS commands this plugin uses timeout. When that happens,
the plugin currently fails out and the user must run the deploy again.
That is bad for automated environments where we want to run deploys
seamlessly. So let's add our own maxRetries to the AWS config object we
get from serverless and make sure that we retry throttled commands
automatically.